Projectiles V2
Good news! Got it to trigger better. Still thinking using another device to aim would probably be best, but it's good for now. Also created some obstacles to test it out. It doesn't trigger on the normal planets now, as it's only triggered by actors. This will make it easier to determine what can and cannot be interacted with.

Bad news. It might just be my computer or how much I've been working on this, but it's laggy for some reason? Hoping it's just on my end. Otherwise we have something to investigate.