Dash Action
Created a basic dash action that goes super fast. We can adjust duration and speed through testing feel.

Also made it so bullets expire on their own for a duration that almost perfectly lines up with the distance and speed of being shot at the slowest speed from one end of the map to the other. This will keep up performance and may sure we don't have any stray assets bogging down the system.