Death Screen Rough Draft
Basic Death Screen was made. No aesthetics applied, only function. When main menu is pressed, quits the game. When start over is pressed, restarts the level.

Does have small bug that you need to click twice when mouse appears. Might not be applicable in the actually built game, but I'm unsure. Also you need to click on the normal top-down screen to make it disappear. Unsure of how to fix this at the moment.