Map Generation
Created a basic script that automatically generates a map, including obstacles and Bonkers. This number is randomized but can be adjusted, and generation of both things can exist in maps where pre-determined objects have already been inserted. I left several placeholder planets to showcase this.

Currently when you activate 'start over' on the death screen, it automatically regenerates a map. We will need to come up with a means to save information so that players can return to where they want to start over rather then a brand new map. However, for now, it's a good way to show the variations within the generation.