Asteroid Hazard Rough Draft
A floating ring has been created, where a series of destructible and hazardous asteroids are spawned. They float around and follow the path of the ring.

Later on, this can be added to the level spawner once we determine how the map spawner should prioritize. Caution: The more we introduce to the spawner, the more dangerous the 'check and re-place' loop becomes. Will likely need a failsafe for this to prevent infinite loops.

Additionally, this hazard has a bug. When the player is hit with an asteroid, they hit the surface and then phase through it. Currently unsure how to fix this.