Junkyard Hazard Rough Draft
Generated in the center of the map is the test of a Junkyard hazard. These are small physics objects of various sizes that are intended to be destroyed with ship projectile and also inflict small amounts of damage when they hit the ship. To control Junkyard testing, walls have been increased in height, now only show in the minimap.

Currently doesn't inflict damage if the objects are too small. Unsure of how to fix this at the moment. In the same vein, also has trouble registering projectile.

This is more of a 'check in' of progress. It's a lead, not a completed task.

Additionally, damage tracking has been made more uniform and generalized so the same method is used for all damageable entities.