Map Generation V2
All hazards have been added to the map generation. I've made a percentage and chance-based system that dictates the likelihood of certain hazards appearing together or combat within the certain hazards.

Currently, Asteroids and Junkyards spawn around a specific dead planet if the two are together. However, it's unclear how it's picking which planet. Code says it's the biggest structure, but that's not the case.  It seems to be random. Odd bug. Also, asteroids currently decimate anything in their path upon spawning. On top of that, it was having severe issues if I tried to prevent it from spawning in with collision to itself or anything else, so I had to deactivate it to get the prototype to work. Will need to return to this later. to root out the problems origins.

Specific proportions, combinations, design elements, and etc can be determined through playtesting.