Background Parallax
Created a basic background and a parallax movement effect. That way, it feels much less disorienting to move around and also see other objects moving around. This image is just a filler at the moment and will be replaced with something more visually in-line with the game at a later time.