Soft Boundaries
Rather then strict walls that prevent the player from leaving, the player is allowed to temporarily go out of bounds. Once they go past that point, however, a widget appears with a countdown. For every second spent before 0, the players movement slows down until a near complete stop. This prevents the player from dashing off into the sunset or off the map. They can turn around and, once exiting the area, their movement returns to normal. If the countdown hits 0, the level is quit. This will later become the loader for the map view level.

One major issue with this is the way it effects the AI. I have a hunch as to what's causing it and how to fix it, however. When you leave the map and go 'out of bounds', all of the AIs freeze. Even when you return to the map, they stay frozen. I believe this is because the enemies do not have a secondary objective. As a result, they cannot do their primary and cease to function. Creating an idle roam will most likely fix this. For now, they freeze.