Still Witness Rough Draft
Basic enemy type of stationary witnesses have been made. To begin, line tracing is used to create a line-of-site indicator. When the player gets in range, a sniper-like line is formed. It then starts a mental countdown that we can create more obvious indicator for later. Once that countdown hits a certain number of seconds passed, it shoots a beam that deals damage to anything it hits. Once it shoots, it can't move until the beam de-activates, allowing the player to move out of site or maneuver if need be.

I also did basic implementation of Witnesses spawning into the map just as an example. This can be adjusted later.

Ideally, these beams would effect other Witnesses. Currently, they do not. I'm not sure why Bonkers that spawn in have different names, but Witnesses don't seem to despite having the same processes. Additionally, Witnesses spawning in on the map has a strange phenomenon where it makes the ships projectile speed unable to scale with the speed of the ship. It doesn't use the projectile speed variable at all from what I can tell, so I have no idea why. Will need to investigate!