Bomb Junk
Scattered amongst the junk in the junkyard are now pink-ish jump pieces. These are unstable, explosive junk. Once the ship makes contact with it, it begins to brighten and a small timer is triggered. As it runs out, it explodes, pushing back physics objects and other enemies. It inflicts damage onto anything within that explosion range, including objects, enemies, and the ship.