Still Witness V2
Still Witnesses spawn more in the middle of the map rather then the edges to increase the likelihood of interacting with them. This location and range can be adjusted on the map later on. The number of Still Witnesses has been increased as well, making them more of a threat. The range of the beam has been doubled to make it harder to avoid, and also allow the player to hit other enemies with the attack easier.