Enemy Update
Enemies now have a placeholder 'light' that shows what they can see in front of them. Both Still Witnesses and Bonkers have idle animations, allowing them to behave naturally in the map when the player isn't around. This makes Bonkers less of an immediate threat, so I also increased the average numbers on the map to make them more pressing.

This will also make Bonkers and Witnesses more susceptible to visual obstruction to allow the player to escape.

The idle animation for the Still Witnesses doesn't work exactly as intended. Can't currently figure out how to do this, but animation is paused when the Witness sees the ship, and resumed when it leaves its sight. Unfortunately, because it's resuming it doesn't use the last position the Witness was in, causing it to skip. Need a better way to do this.