Game Seed + Tile Seeds
Text of O (obstacle), P (planet), or C (combat) now appears on all tiles that are not empty. This values are also tied to the tiles using structs, which can be used when the text is no longer needed.

The shown text was generated by utilizing the game seed, which can be set manually currently. The value will be randomly generated and saved later when the main menu and a save file system is implemented.

Additionally, each tile has been given a seed number. This allows each tile to be generated independently when the players loads in, with the game patching them together via the overall game seed. Currently the top-down generator doesn't utilize these, as this is just a set up. Tiles do not trigger.

Additionally, the visibility of the tile events has been limited to visible tiles, allowing for playtesting of travel. How close together and how often certain tiles appear can be determined through playtesting.