Obstacle vs Combat Tiles
The top-down level generator method has been altered to change between combat and obstacles depending on the given type made by the tiles themselves. Almost all level of randomization has been transferred to utilizing the game and tile seeds, though this will need to be reviewed thoroughly to make sure this is the case.

Additionally, to increase efficiency, interfaces are beginning to be implemented instead of casting where possible.