Inventory System
Basic inventory system has been implimented in the Cargo Level. This view only includes the cargo inventory. Additionally, there is currently no connection between actually collected resources and what is in the inventory.

However, it does properly take resources and stack them within the inventory. Those that cannot stack aren't stacked. Those that are, have a max stack size and are limited by it. It also has a max inventory size implemented at the moment. Only current resources in the inventory for testing are metal and medicine, but this can easily be expanded upon.

Additionally, variables and functions can be referenced across the project using the game instance, interfaces, and event dispatchers. This will make it easier to implement collection in the Top Down map to then be processed in the Cargo view.