Inventory V2
Inventory now updates using the AddToInven function, which can be referenced throughout the entire game. These update the location and amount of items in the inventory without the player being in the inventory at the time. This also uses the Game Instance, which means this automatically saves whenever you get an object even if you don't access your inventory over a period of time.

This does not have current drag-and-drop implementation, nor adaptation for possible alternatively sized, unstackable objects.