Narrative System
Basic dialog capabilities have been expanded to a full narrative system. Conversations move forward and cannot be repeated. Additionally, morale can effect what conversation is selected automatically.

Additional factors to expand on what adds or removes conversations from the narrative path are yet to be developed, as it will likely be a case-by-case basis of development. This would be done best with independent functions.

A simple interface has been created to allow the player to choose who they speak to, which in turn selects what narrative they're progressing. Unsure as of yet how to go about two or even three person conversations, as this will need to be developed and expanded later.