Commerce System Rough Draft
You can now by and sell resources at the marketplace. This increases or decreases your inventory accordingly, as well as increases or decreases the 'scrap'.

Currently this is a very buggy system, but foundation functions work. First issue, if there is no medicine in the players inventory, it will struggle to use it in the layout of the marketplace. Additionally, one additional resource always appears in the right resource grabber section of the marketplace. Unsure why this is or how to make it not happen. Lastly, when you buy resources from the marketplace and add them to your inventory, it will also increase the number of items you pick up from the marketplace at a time. For example, if you have five metal, you will pick up five metal from the marketplace. Seems to be adding to both inventories somehow.

Last but not least, the scrap decrease function has not be fully set up yet, as it's hard to tell if it's actually working due to the large volume of strange bugs weirdly influencing it.